+++ lib/moon/components/toast.ex
@@ -11,6 +11,7 @@ defmodule Moon.Components.Toast do
   alias Moon.Components.Link
 
   prop id, :string, required: true
+  prop testid, :string
   prop message, :string, required: true
   prop variant, :string, values: ~w(error warning info success)
   prop closeable, :boolean, default: true
@@ -22,7 +23,13 @@ defmodule Moon.Components.Toast do
   defmodule Message do
     @moduledoc false
 
-    defstruct id: nil, message: nil, variant: nil, closeable: true, link_text: nil, link_href: nil
+    defstruct id: nil,
+              testid: nil,
+              message: nil,
+              variant: nil,
+              closeable: true,
+              link_text: nil,
+              link_href: nil
   end
 
   def render(assigns) do
@@ -55,6 +62,7 @@ defmodule Moon.Components.Toast do
           class="inline-flex rounded-md hover:text-gray-500 focus:outline-none"
           :on-click={@on_close}
           phx-value-id={@id}
+          data-testid={"#{@testid}-close-button"}
         >
           <span class="sr-only">Close</span>
           <IconErrorCircle class="w-5 h-5" font_size="1.25rem" color="trunks-100" />