+++ lib/moon_web/pages/assets/logos_page.ex
@@ -28,7 +28,7 @@ defmodule MoonWeb.Pages.Assets.LogosPage do
 
   def render(assigns) do
     ~F"""
-    <Page theme_name={@theme_name} active_page={@active_page} breadcrumbs={@breadcrumbs}>
+    <Page {=@theme_name} {=@active_page} {=@breadcrumbs} {=@direction}>
       <TopToDown>
         <Heading size={56} class="mb-4">Logos</Heading>
 
@@ -262,6 +262,30 @@ defmodule MoonWeb.Pages.Assets.LogosPage do
 
         <ExampleAndCode id="logo_20" class="mt-4">
           <:example>
+            <Logo name="logo_moneyball_io_full" font_size="10rem" />
+          </:example>
+
+          <:code>
+            <#CodePreview>
+              <Logo name="logo_moneyball_io_full" font_size="10rem" />
+            </#CodePreview>
+          </:code>
+        </ExampleAndCode>
+
+        <ExampleAndCode id="logo_21" class="mt-4">
+          <:example>
+            <Logo name="logo_moneyball_io_short" font_size="10rem" />
+          </:example>
+
+          <:code>
+            <#CodePreview>
+              <Logo name="logo_moneyball_io_short" font_size="10rem" />
+            </#CodePreview>
+          </:code>
+        </ExampleAndCode>
+
+        <ExampleAndCode id="logo_22" class="mt-4">
+          <:example>
             <Logo name="logo_moneyball" font_size="10rem" />
           </:example>
 
@@ -272,7 +296,7 @@ defmodule MoonWeb.Pages.Assets.LogosPage do
           </:code>
         </ExampleAndCode>
 
-        <ExampleAndCode id="logo_21" class="mt-4">
+        <ExampleAndCode id="logo_23" class="mt-4">
           <:example>
             <Logo name="logo_moon_design_short" font_size="10rem" />
           </:example>
@@ -284,7 +308,7 @@ defmodule MoonWeb.Pages.Assets.LogosPage do
           </:code>
         </ExampleAndCode>
 
-        <ExampleAndCode id="logo_22" class="mt-4">
+        <ExampleAndCode id="logo_24" class="mt-4">
           <:example>
             <Logo name="logo_moon_design" font_size="10rem" />
           </:example>
@@ -296,7 +320,7 @@ defmodule MoonWeb.Pages.Assets.LogosPage do
           </:code>
         </ExampleAndCode>
 
-        <ExampleAndCode id="logo_23" class="mt-4">
+        <ExampleAndCode id="logo_25" class="mt-4">
           <:example>
             <Logo name="logo_paxful" font_size="10rem" />
           </:example>
@@ -308,7 +332,7 @@ defmodule MoonWeb.Pages.Assets.LogosPage do
           </:code>
         </ExampleAndCode>
 
-        <ExampleAndCode id="logo_24" class="mt-4">
+        <ExampleAndCode id="logo_26" class="mt-4">
           <:example>
             <Logo name="logo_powered_by_hub88" font_size="10rem" />
           </:example>
@@ -320,7 +344,7 @@ defmodule MoonWeb.Pages.Assets.LogosPage do
           </:code>
         </ExampleAndCode>
 
-        <ExampleAndCode id="logo_25" class="mt-4">
+        <ExampleAndCode id="logo_27" class="mt-4">
           <:example>
             <Logo name="logo_slots_full" font_size="10rem" />
           </:example>
@@ -332,7 +356,7 @@ defmodule MoonWeb.Pages.Assets.LogosPage do
           </:code>
         </ExampleAndCode>
 
-        <ExampleAndCode id="logo_26" class="mt-4">
+        <ExampleAndCode id="logo_28" class="mt-4">
           <:example>
             <Logo name="logo_slots_short" font_size="10rem" />
           </:example>
@@ -344,7 +368,7 @@ defmodule MoonWeb.Pages.Assets.LogosPage do
           </:code>
         </ExampleAndCode>
 
-        <ExampleAndCode id="logo_27" class="mt-4">
+        <ExampleAndCode id="logo_29" class="mt-4">
           <:example>
             <Logo name="logo_sportsbet_full" font_size="10rem" />
           </:example>
@@ -356,7 +380,7 @@ defmodule MoonWeb.Pages.Assets.LogosPage do
           </:code>
         </ExampleAndCode>
 
-        <ExampleAndCode id="logo_28" class="mt-4">
+        <ExampleAndCode id="logo_30" class="mt-4">
           <:example>
             <Logo name="logo_sportsbet_short" font_size="10rem" />
           </:example>
@@ -368,7 +392,7 @@ defmodule MoonWeb.Pages.Assets.LogosPage do
           </:code>
         </ExampleAndCode>
 
-        <ExampleAndCode id="logo_29" class="mt-4">
+        <ExampleAndCode id="logo_31" class="mt-4">
           <:example>
             <Logo name="logo_sportsbet_sponsor" font_size="10rem" />
           </:example>
@@ -380,7 +404,7 @@ defmodule MoonWeb.Pages.Assets.LogosPage do
           </:code>
         </ExampleAndCode>
 
-        <ExampleAndCode id="logo_30" class="mt-4">
+        <ExampleAndCode id="logo_32" class="mt-4">
           <:example>
             <Logo name="logo_sportsbet" font_size="10rem" />
           </:example>
@@ -392,7 +416,7 @@ defmodule MoonWeb.Pages.Assets.LogosPage do
           </:code>
         </ExampleAndCode>
 
-        <ExampleAndCode id="logo_31" class="mt-4">
+        <ExampleAndCode id="logo_33" class="mt-4">
           <:example>
             <Logo name="logo_utorg" font_size="10rem" />
           </:example>
@@ -404,7 +428,7 @@ defmodule MoonWeb.Pages.Assets.LogosPage do
           </:code>
         </ExampleAndCode>
 
-        <ExampleAndCode id="logo_32" class="mt-4">
+        <ExampleAndCode id="logo_34" class="mt-4">
           <:example>
             <Logo name="moon_pay" font_size="10rem" />
           </:example>